feat: terminal injection API, df/meminfo commands, process Ctrl+C, sound/scanf fixes
- Add IPO_SYSCALL_TERMINAL_INPUT: inject text into terminal prompt with optional auto-execute

- Add term_ctl app (execute / -w write-only) and exec_caller app (ipo_exec API)

- Add builtin df/diskinfo and meminfo/free commands

- Fix Ctrl+C during scanf: return EINTR, print 'Application stopped'

- Remove sound frequency limits, fix scanf whitespace/scanset handling

- Add process ps/kill/killall builtins, kmalloc_get_stats

+++ lib/h/driver/sound.h
@@ -12,7 +12,7 @@ void sound_init(void);
  * Play sound at specified frequency (Hz)
  * @param frequency Frequency in Hz (440 = A4, typical range 37-20000 Hz)
  */
-void sound_play(uint16_t frequency);
+void sound_play(uint32_t frequency);
 
 /**
  * Stop playing sound
@@ -24,7 +24,7 @@ void sound_stop(void);
  * @param frequency Frequency in Hz
  * @param duration Duration in milliseconds (approximate)
  */
-void sound_beep(uint16_t frequency, uint16_t duration);
+void sound_beep(uint32_t frequency, uint32_t duration);
 
 /* Define musical notes with their corresponding frequencies (in Hz) */
 #define NOTE_REST 0
